requirements.txt
# 'pip install -e .' (editable, dev files) would automatically install these dependences
psycopg2==2.8.5
Scrapy==2.0.1
Django==3.0.5
Pillow==7.0.0
scrapy-djangoitem==1.1.1
graypy==2.1.0
scrapy-crawlera==1.7.0
python-scrapyd-api==2.1.2
scrapyd-client==1.1.0
djangorestframework==3.11.0
requests==2.23.0
treq==20.4.1